+++ index.js
@@ -141,17 +141,19 @@ class Site {
 
     // Add domain aliases support (add 302 redirects for any domains
     // defined as aliases so that the URL is rewritten). There is always
-    // at least one alias (the www. subdomain).
-    const mainHostname = os.hostname()
-    this.app.use((request, response, next) => {
-      const requestedHost = request.header('host')
-      if (requestedHost === mainHostname) {
-        next()
-      } else {
-        console.log(` 👉 [Site.js] Redirecting alias ${requestedHost} to main hostname ${mainHostname}.`)
-        response.redirect(`https://${mainHostname}${request.path}`)
-      }
-    })
+    // at least one alias (the www. subdomain) for global servers.
+    if (this.global) {
+      const mainHostname = os.hostname()
+      this.app.use((request, response, next) => {
+        const requestedHost = request.header('host')
+        if (requestedHost === mainHostname) {
+          next()
+        } else {
+          console.log(` 👉 [Site.js] Redirecting alias ${requestedHost} to main hostname ${mainHostname}.`)
+          response.redirect(`https://${mainHostname}${request.path}`)
+        }
+      })
+    }
 
     // Statistics view (displays anonymous, ephemeral statistics)
     this.app.get(this.stats.route, this.stats.view)