+++ examples/wss-routes/.dynamic/.wss/echo.js
@@ -1,0 +1,6 @@
+// Basic echo.
+module.exports = (webSocket, request) => {
+  webSocket.on('message', (data) => {
+    webSocket.send(data)
+  })
+}