+++ test/site-dynamic-dotjs-separate-get-and-post/.dynamic/.post/file-name-as-route-name.js
@@ -1,0 +1,3 @@
+module.exports = (request, response) => {
+  response.type('text').end('POST /file-name-as-route-name')
+}