+++ test/site-dynamic-dotjs-separate-https-and-wss-and-separate-get-and-post/.dynamic/.wss/file-name-as-route-name.js
@@ -1,0 +1,5 @@
+module.exports = (webSocket, request) => {
+  webSocket.on('message', (data) => {
+    webSocket.send(`/file-name-as-route-name ${data}`)
+  })
+}