+++ test/site/.dynamic/deep/deeper/index.js
@@ -1,0 +1,3 @@
+module.exports = (request, response) => {
+  response.type('html').status(200).end('<h1>Deeper index</h1>')
+}